@@ -12,45 +12,38 @@
     "lint": "tslint ./src/**/*.ts -t verbose",
     "e2e": "ng e2e"
   },
-  "keywords": [],
-  "author": "",
-  "license": "MIT",
   "dependencies": {
-    "@angular/animations": "^15.2.0",
-    "@angular/common": "^15.2.0",
-    "@angular/compiler": "^15.2.0",
-    "@angular/core": "^15.2.0",
-    "@angular/forms": "^15.2.0",
-    "@angular/localize": "^15.2.0",
-    "@angular/platform-browser": "^15.2.0",
-    "@angular/platform-browser-dynamic": "^15.2.0",
-    "@angular/router": "^15.2.0",
-    "@angular/upgrade": "^15.2.0",
-    "@fortawesome/angular-fontawesome": "^0.12.1",
-    "@fortawesome/fontawesome-common-types": "0.2.30",
-    "@fortawesome/fontawesome-free": "5.14.0",
-    "@fortawesome/fontawesome-svg-core": "1.2.30",
-    "@fortawesome/free-solid-svg-icons": "5.14.0",
-    "@ng-bootstrap/ng-bootstrap": "^14.0.1",
-    "angular-in-memory-web-api": "~0.11.0",
-    "bootstrap": "^4.5.3",
-    "file-saver": "^1.3.8",
-    "ngx-image-cropper": "^5.0.0",
+    "@angular/animations": "^17.0.0",
+    "@angular/common": "^17.0.0",
+    "@angular/compiler": "^17.0.0",
+    "@angular/core": "^17.0.0",
+    "@angular/forms": "^17.0.0",
+    "@angular/platform-browser": "^17.0.0",
+    "@angular/platform-browser-dynamic": "^17.0.0",
+    "@angular/router": "^17.0.0",
+    "@fortawesome/angular-fontawesome": "^0.14.0",
+    "@fortawesome/fontawesome-svg-core": "^6.4.0",
+    "@fortawesome/free-solid-svg-icons": "^6.4.0",
+    "@ng-bootstrap/ng-bootstrap": "^16.0.0",
+    "bootstrap": "^5.3.2",
+    "file-saver": "^2.0.5",
+    "ngx-image-cropper": "^8.0.0",
     "qrcode": "^1.4.4",
-    "rxjs": "~6.5.4",
+    "rxjs": "^7.8.0",
     "xlsx": "^0.18.5",
-    "zone.js": "~0.12.0"
+    "zone.js": "~0.14.6"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.0",
-    "@angular/cli": "^15.2.0",
-    "@angular/compiler-cli": "^15.0.0",
-    "@types/node": "^12.19.9",
+    "@angular-devkit/build-angular": "^17.0.0",
+    "@angular/cli": "^17.0.0",
+    "@angular/compiler-cli": "^17.0.0",
+    "@types/file-saver": "^2.0.7",
+    "@types/node": "^17.0.0",
+    "@types/qrcode": "^1.5.5",
     "lodash": "^4.17.21",
     "protractor": "^7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "4.8.2"
-  },
-  "repository": {}
+    "typescript": "<5.3.0"
+  }
 }
