@@ -10,8 +10,6 @@ import java.util.List;
 @Entity
 @Data
 @NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode
 @Table(name = "STAND")
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Stand {
@@ -23,12 +21,6 @@ public class Stand {
     @Column(name = "NOM", nullable = false)
     private String nom;
 
-    @Column(name = "DESCRIPTION")
-    private String description;
-
-    @Column(name = "BULLE")
-    private String bulle;
-
     @Column(name = "ORDRE")
     private Integer ordre;
 
@@ -36,14 +28,15 @@ public class Stand {
     private Integer type;
 
     @JsonIgnore
+    @ManyToOne(fetch = FetchType.EAGER)
+    @ToString.Exclude
+    private Evenement evenement;
+
     @OneToMany(mappedBy="stand", orphanRemoval=true, fetch = FetchType.EAGER)
+    @JsonIgnoreProperties({"stand"})
     @ToString.Exclude
     private List<Croisement> croisements;
 
-    @JsonIgnore
-    @ManyToOne(fetch = FetchType.EAGER)
-    @JsonIgnoreProperties({"affiche", "sitepersologo"})
-    @ToString.Exclude
-    private Evenement evenement;
+
 
 }
