@@ -2,7 +2,8 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule} from "@angular/forms";
 import { AppComponent } from './app.component';
-import { CoreModule } from "./core/core.module";
+import { CoreModule } from './core/core.module';
+
 
 
 @NgModule({
